CA-299944: Proxy requests to updates
Rather than going over http over the network to fetch the RPMs and yum metadata,
connect to the local xapi and have the requests forwarded over stunnel.

This behaviour is activated by calling the (undocumented) Pool_update.attach
API with the new parameter 'use_localhost_proxy' set to true. The absence of
the parameter is treated as 'false' and xapi will fall back to the old behaviour.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>